Move uninstall application button
